+++ crates/sdk/src/js/instructions/create_shift.rs
@@ -21,11 +21,11 @@ pub struct CreateShiftParamsJs {
     #[serde(default)]
     pub receiver: Option<StringPubkey>,
     #[serde(default)]
-    pub from_market_token_amount: u64,
+    pub from_market_token_amount: Option<u128>,
     #[serde(default)]
-    pub min_to_market_token_amount: u64,
+    pub min_to_market_token_amount: Option<u128>,
     #[serde(default)]
-    pub skip_to_market_token_ata_creation: bool,
+    pub skip_to_market_token_ata_creation: Option<bool>,
 }
 
 #[derive(Debug, Serialize, Deserialize, Tsify)]
@@ -61,7 +61,6 @@ pub fn create_shifts_builder(
     let mut groups: Vec<AtomicGroup> = Vec::with_capacity(shifts.len());
 
     for params in shifts.into_iter() {
-        tokens.insert(params.from_market_token);
         tokens.insert(params.to_market_token);
 
         let program = options.program.clone().unwrap_or_default();
@@ -70,9 +69,21 @@ pub fn create_shifts_builder(
             .payer(options.payer)
             .from_market_token(params.from_market_token)
             .to_market_token(params.to_market_token)
-            .from_market_token_amount(params.from_market_token_amount)
-            .min_to_market_token_amount(params.min_to_market_token_amount)
-            .skip_to_market_token_ata_creation(params.skip_to_market_token_ata_creation);
+            .from_market_token_amount(
+                params
+                    .from_market_token_amount
+                    .unwrap_or_default()
+                    .try_into()?,
+            )
+            .min_to_market_token_amount(
+                params
+                    .min_to_market_token_amount
+                    .unwrap_or_default()
+                    .try_into()?,
+            )
+            .skip_to_market_token_ata_creation(
+                params.skip_to_market_token_ata_creation.unwrap_or_default(),
+            );
 
         let built = if let Some(r) = params.receiver {
             builder.receiver(r).build()