+++ crates/sdk/src/simulation/deposit.rs
@@ -23,7 +23,7 @@ impl DepositSimulationOutput {
     }
 
     /// Returns short swap reports.
-    pub fn short_swap(&self) -> &[SwapReport<u128, i128>] {
+    pub fn short_swaps(&self) -> &[SwapReport<u128, i128>] {
         &self.short_swaps
     }