+++ crates/sdk/src/simulation/glv_deposit.rs
@@ -26,7 +26,7 @@ impl GlvDepositSimulationOutput {
     }
 
     /// Returns short swap reports.
-    pub fn short_swap(&self) -> &[SwapReport<u128, i128>] {
+    pub fn short_swaps(&self) -> &[SwapReport<u128, i128>] {
         &self.short_swaps
     }