+++ crates/sdk/src/simulation/mod.rs
@@ -7,14 +7,16 @@ pub mod order;
 /// Deposit simulation.
 pub mod deposit;
 
+/// Withdrawal simulation.
+pub mod withdrawal;
+
+/// Shift simulation.
+pub mod shift;
+
 /// GLV deposit simulation.
 pub mod glv_deposit;
 
-/// Liquidity (GM) exact simulation.
-pub mod lp;
+/// GLV withdrawal simulation.
+pub mod glv_withdrawal;
 
-pub use lp::{
-    simulate_gm_deposit_exact, simulate_gm_withdrawal_exact, GmDepositExactOutput,
-    GmWithdrawalExactOutput,
-};
 pub use simulator::{SimulationOptions, Simulator, SwapOutput, TokenState};